* work on cast.types
